# internal/server/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gemba HTTP API",
    "version": "0.1.0",
    "description": "Wire schema for the gemba server's `/api/*` surface (gm-e4.2). The SPA at `web/` consumes a typed client generated from this document; the source of truth lives in `docs/api/openapi.json` and is served at runtime under `/api/openapi.json`.\n\nThis MVP covers the read endpoints the SPA depends on day-to-day: health, version, config, repositories, work items, sprints, agents, sessions, escalations, capabilities, and the new-project conversational flow. The remaining `/api/*` surface (consults, walks, planner, perspectives) is acknowledged in `paths` with summary-only stubs and will gain full request/response schemas in follow-up beads.",
    "contact": {
      "name": "GembaCore/gemba-core",
      "url": "https://github.com/GembaCore/gemba-core"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (the gemba server is the SPA host)."
    }
  ],
...
  "paths": {
...
    "/api/work-items": {
      "get": {
        "tags": ["work"],
        "summary": "List work items via the bound WorkPlane adaptor.",
        "operationId": "listWorkItems",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by WorkItem kind (e.g. `epic`, `task`).",
            "schema": { "type": "string" }
          },
          {
            "name": "state_category",
            "in": "query",
            "description": "Filter by normalized state category.",
            "schema": { "$ref": "#/components/schemas/StateCategory" }
          },
          {
            "name": "label",
            "in": "query",
            "description": "Filter by label (repeatable).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "WorkItem list envelope.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListWorkItemsEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
    "/api/work-items/{id}": {
      "get": {
        "tags": ["work"],
        "summary": "Fetch one WorkItem by id.",
        "operationId": "getWorkItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Workspace-prefixed bead id (e.g. `gemba/gm-e4.2`).",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The work item.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkItem" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "503": { "$ref": "#/components/responses/AdaptorUnavailable" }
        }
      }
    },
...
    "/api/v1/projects/bind": {
      "post": {
        "tags": ["work"],
        "summary": "Bind a partially-bootstrapped project to a git repo (gm-root.17.14). `mode=\"create\"` runs `git init` at the beads-DB directory and writes `.gemba/workspace.toml` in place. `mode=\"navigate\"` copies the beads DB into an existing git repo and writes `.gemba/workspace.toml` there. Returns the resolved `ProjectEntry` with `kind=\"complete\"` so the SPA can switch the active workspace immediately.",
        "operationId": "bindProject",
        "parameters": [
          {
            "name": "X-GEMBA-Confirm",
            "in": "header",
            "required": true,
            "description": "Idempotency nonce. Replays of the same nonce return the cached response.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BindProjectRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bind succeeded. Body is the now-complete `ProjectEntry`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEntry" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": {
            "description": "Conflict — `mode=\"navigate\"` and the target repo already contains a `.beads/` directory the server refuses to clobber (`error` = `beads_dir_exists`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server-side failure during git_init / beads_copy / workspace_toml write.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Malformed request — invalid query parameters or body.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "AdaptorUnavailable": {
        "description": "Adaptor is not registered or is degraded; the SPA shows a banner and keeps cached data visible.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Shared error envelope (gm-faz). `error` is a stable token the SPA branches on; `message` is human-readable and may be surfaced verbatim.",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "examples": ["adaptor_degraded", "validation", "not_found"] },
          "message": { "type": "string" }
        }
      },
...
      "StateCategory": {
        "type": "string",
        "enum": ["backlog", "unstarted", "staged", "started", "completed", "canceled"]
      },
      "AgentRef": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" }
        }
      },
      "WorkItem": {
        "type": "object",
        "description": "Subset of the WorkItem fields most commonly read by the SPA. The full shape (relationships, evidence, dod, derived signals, custom fields) is the source of truth in `core/types.go` and `web/src/types/core.gen.ts`; this schema documents the wire-stable subset for a typed client. `additionalProperties: true` keeps clients forward-compatible with fields landing in follow-up beads.",
        "additionalProperties": true,
        "required": ["id", "kind", "title", "status", "state_category", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string", "examples": ["epic", "task", "feature", "decision"] },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string" },
          "state_category": { "$ref": "#/components/schemas/StateCategory" },
          "priority": { "type": ["integer", "null"], "minimum": 0, "maximum": 4 },
          "owner": { "$ref": "#/components/schemas/AgentRef" },
          "assignee": { "$ref": "#/components/schemas/AgentRef" },
          "labels": { "type": "array", "items": { "type": "string" } },
          "primary_repository_id": { "type": "string" },
          "repository_ids": { "type": "array", "items": { "type": "string" } },
          "sprint_id": { "type": ["string", "null"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListWorkItemsEnvelope": {
        "type": "object",
        "required": ["items", "total"],
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/WorkItem" }
          },
          "total": { "type": "integer" }
        }
      },
...
      "ProjectKind": {
        "type": "string",
        "description": "Setup-state classification for a discovered project (gm-root.17.14). `complete` — has a beads DB, a `.gemba/workspace.toml`, and is inside a git repo; the picker switches to it directly. `needs_workspace` — has a beads DB but no `.gemba/workspace.toml`. `needs_repo` — has a `.gemba/workspace.toml` but is not inside a git repo. Open string at the wire so future kinds don't break clients.",
        "enum": ["complete", "needs_workspace", "needs_repo"]
      },
      "ProjectEntry": {
        "type": "object",
        "description": "One discovered project. `name` is the directory basename; `path` is the absolute disk path; `active` is true for the currently-selected project (in-process server state; resets on server restart); `kind` classifies the entry's setup state.",
        "required": ["name", "path", "kind"],
        "properties": {
          "name": { "type": "string" },
          "path": { "type": "string" },
          "active": { "type": "boolean" },
          "kind": { "$ref": "#/components/schemas/ProjectKind" }
        }
      },
...
      "BindProjectRequest": {
        "type": "object",
        "description": "Body for `POST /api/v1/projects/bind`. `mode=\"create\"` runs `git init` at `beads_db_path` (which must equal `target_repo_path`) and writes `.gemba/workspace.toml` in place. `mode=\"navigate\"` copies the beads DB at `beads_db_path/.beads/` into `target_repo_path/.beads/` and writes `.gemba/workspace.toml` at `target_repo_path` — `target_repo_path` must already be a git working tree.",
        "required": ["beads_db_path", "target_repo_path", "mode"],
        "properties": {
          "beads_db_path": {
            "type": "string",
            "description": "Absolute path to the directory holding the beads DB (the parent of the `.beads/` subdir)."
          },
          "target_repo_path": {
            "type": "string",
            "description": "Absolute path to the directory to bind to. Equals `beads_db_path` when `mode=\"create\"`."
          },
          "mode": {
            "type": "string",
            "enum": ["create", "navigate"]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token auth (gm-e5.1). Required when the server boots with `--auth=token`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gemba_session",
        "description": "Signed session cookie issued by `POST /api/auth/login` after a bearer exchange."
      }
    }
  },
  "security": [
    { "bearerToken": [] },
    { "sessionCookie": [] },
    {}
  ]
}